Escape more template parameters
Grouper was vulnerable to reflected XSS via the search field when
showing search results.  Fix that by escaping the search query on
the results page, and escape various other parameters that might
conceivably be under user control.  (A more comprehensive audit in
the future is probably indicated.)

Names of user passwords were not previously restricted to exclude
various HTML metacharacters.  Do that as well.