Be consistent about using current_user attribute
With Tornado, if you define a get_current_user method, it will use
this to populate current_user on demand and cache the result.  We
were using that in most places, but calling get_current_user
directly in some places, which was confusing.  Use the current_user
attribute everywhere, and ensure that everything works properly if
get_current_user returns None.