Enable autoescape in Jinja2 templates
Turn on autoescaping for all HTML templates.  Stop escaping in the
text email templates where this is meaningless anyway.  Explicitly
mark as safe the one place we inject raw HTML (via xsrf_form from
Tornado).

Remove all the old explicit escaping in the Jinja2 templates, since
it would now be confusing (and is just clutter).

Rework the help_for macro so that it doesn't inject its results
into the middle of a tag, which forces escaping and confuses Jinja2.

Rename all email templates to end in .html or .txt for easier
configuration of auto-escaping, except for the subject line
templates that still end in .tmpl.

Add a simple test for searching in the frontend, and use it to test
that escaping is happening.  The test itself has been tested by
marking the search input as safe and confirming the test then fails.
Add class markings to the search results page so that we can more
easily test with Selenium.

Import some additional models in SetupTest so that it can be used
for additional frontend pages.  This is a temporary hack until schema
creation is moved into a proper Repository object.