Move database URL handling into Settings
Rather than using a separate utility function to interpret Settings
and maintain a separate global cache, move database_url to a
property on Settings and the logic for running a database_source
command to private methods.  This simplifies a lot of callers.

In the API and frontend servers, properly override the database URL
if a URL was given on the command line.  This prevents the server
from later running a database_source command and overriding a URL
explicitly given via a command-line argument.

Add a test for Settings behavior, including this new code.