Rework grouper-ctl sync_db and start on schema service
Move initialization of the schema into a usecase, and rework the
sync_db grouper-ctl command as a new-style CtlCommand.  Not all of
its functions have moved to usecases yet, but it's a start.

This uncovered a problem with grouper-ctl testing: we were passing
in the config/dev.yaml configuration file, but it uses a different
database URL than tests.  Fix this by adding a -d flag to
grouper-ctl to specify a database URL, consistent with the API
and frontend servers, and have the test suite pass that in to
override the database URL.  This required adding a new parameter
to the legacy run_main test method.

Reorganize the test suite to separate sync_db into its own test and
test more of its functionality.