Revoke inactive permission grants on disable
To fully establish the invariant that disabled permissions are not
granted to anything, we have to handle the case of disabled groups.
(Service accounts delete all of their permission grants when
disabled, so are not affected.)  For enabled groups, we require the
person disabling the permission revoke all the grants first, but it
doesn't make sense to ask them to do this for disabled groups, which
are invisible in the UI.

Semantically, it makes sense to silently revoke the permission from
all disabled groups, since their grants are retained only to preserve
the group state if disabling it was a mistake, and removing grants of
a separately-disabled permission is not a mistake.

Do that on disable, and add a test that it happens properly.