Remove caching of database_source results
On further inspection of the code, none of the internal users
benefit from caching the URL.  Either they set the URL once on
startup, or they check to see if the URL has changed once in
each loop, and the loops take at least 60 seconds.  Given that,
there's no need for the caching with random delay, and all that
code and tests can be deleted.  Just call the database_source
command every time the Settings object is asked for the database
URL.