Reinitialize plugins between tests
The PluginProxy object is global and remembers previously-configured
plugin paths, and the grouper-ctl tests initialize plugins.  This
was accidentally not causing test failures because pytest was running
tests in a predictable order that ran all plugin-sensitive tests
before grouper-ctl tests by accident, but a different test framework
that ran the tests in a different order showed this problem.

Reinitialize plugins between tests for now until we can fix this
problem properly by injecting the PluginProxy into classes that need
it.