Further simplify database URL handling
The database URL code was still making the problem too complicated
given that we decided to always run the database_source command.
Simplify further to not cache the results and just run the command
directly and return its stripped output.

Use properties and setters to further streamline the code and make
assigning directly to the timezone attribute work properly.  This
allows us to get rid of the special timezone_object and database_url
attributes and have all code use timezone and database, further
hiding the behind-the-scenes magic.