Start injecting PluginProxy where needed
Start the process of migrating away from a global PluginProxy
object to dependency injection.  Move the main entry point for
loading plugins to a static method on PluginProxy, use it to
create a PluginProxy object in each executable, and (for now)
explicitly set it as the global plugin proxy.  Similarly set
an empty one in test initialization.

Move the plugin loading code out of the __init__.py file to
avoid a circular dependency with grouper.plugin.proxy, and also
because code in __init__.py is confusing.  Now only the legacy
global object handling is there.

PluginProxy objects are created by each executable rather than
by the global factory functions because they take the executable
name as a parameter.  Use hyphens instead of underscores for
grouper-api and grouper-fe (a backward-incompatible change) for
consistency (and because Dropbox internal plugins were rewriting
the underscore back to a hyphen anyway).