Add a separate SessionFactory
Extensively rework how sessions are created and injected to use a
separate SessionFactory that's injected into the RepositoryFactory.
This requires passing in some additional parameters to grouper-ctl
legacy commands until we can convert everything to usecases, and
requires a bit more machinery when building sessions and use cases.

On the plus side, this means that session injection from the test
suite works much more reliably, and allows removal of a bunch of
patching from the grouper-ctl test suite.