Refactor templating into new TemplateEngine classes
Move grouper.fe.template_util into a new BaseTemplateEngine class
that creates the Jinja2 template engine with some standard filters
and globals and can serve as the base class for child classes that
add additional filters and globals.  Create child classes for the
frontend template engine and the email template engine.

Add mypy type annotations to email functions to further test this
change, and clean up resulting latent bugs.  Add asserts in a few
places where users are retrieved from the database and it should be
"impossible" for them to not exist.