Force https for default CDNJS configuration
We should always use HTTPS even if the page is HTTP since
developing with a local proxy will be HTTP and we still don't want
to allow an attacker to mess with the JavaScript.  Change the
default configuration setting.

Fix an invalid reference to one style sheet uncovered by that
change.