Use configuration settings for user_proxy auth header
grouper-ctl user_proxy was hardcoding the default value of the
authentication header, and therefore wouldn't work if it were
changed in the configuration.  Instead, pass the configuration
down into the user proxy code and get the authentication header
value from there.

This requires moving the authentication header from the fe section
to the general section of the configuration.