Rework invariant enforcement for disabled permissions
Rather than having to decide whether to perform a sanity check in
the service, have the permission service always revoke all grants
for a permission when it is disabled.  This ensures the invariant
without having to check and throw exceptions.  Rename the relevant
call to make this clear, and add the logic to revoke service account
grants as well.