Set Content-Security-Policy and Referrer-Policy
Build a Content-Security-Policy header from external resources and
the configured CDNJS prefix and set it on every request, including
some restrictive default settings.  Also set Referrer-Policy to
same-origin on every request.

Unfortunately, set_default_headers is called before initialize in
request handlers, so the template engine had to move to the Tornado
settings to be accessible early enough.  Since we're making that
change anyway, also move the session class and drop the parameters
to initialize.  This makes us somewhat more vulnerable to changes in
the attributes in newer versions of Tornado, but avoids globals or
messing about with subclassing Application.