Avoid role user SQL queries for every group in graph
When constructing the graph, _get_group_tuples was making a call
to is_role_user for every group.  This in turn translates into a
SQL query to pull the user from the database.  However, this method
is only called after the user metadata has been cached in the
graph, which includes role user data for every user.

Avoid the SQL queries by asking the cached graph data if the user
exists and, if so, if it's a role user.