Avoid deferred loads when caching permission metadata
The name to which a permission was granted was loaded via a
relationship from the map to the group, which is a deferred load
and thus was resulting in a SQL query per permission grant.  We
were already joining on Group in the initial query, so just
retrieve the group name as well and include it in the result
tuple, avoiding the need for the additional queries.