Don't grant permissions to auditors_group twice
If sync_db was run a second time on the same database, initializing
the schema would cause a constraint violation because the auditor
permission was traged to the auditors group unconditionally.  Only
grant the permission if the group didn't already exist.

Add a test that sync_db succeeds if run twice on the same database.