Avoid using the template engine in permissions itest
We were initializing the template engine just to be able to call
print_date, but the behavior of that function is easily duplicated
in one line of code and initializing a template engine requires
configuring some dependencies that are a little tedious.  Just
duplicate the expected formatting logic, which is highly unlikely
to change.