Better logging and handling of ownerless expirations
When expiring a group edge, we had been picking a random owner from
the group from which the edge was expiring.  We did this without
checking if that group had an owner, so if it didn't, Grouper would
raise an entirely uninformative StopIteration exception.

Drastically reduce the incidence of this problem by using the user
themselves as the actor ID when expiring a user (since the actor ID
is fairly arbitrary anyway) and, when expiring edges to another group,
using an arbitrary owner of the parent group or an arbitrary owner of
the child group.

When expiring a group edge between two groups, both of which have no
owners, there isn't much we can do given the current design of the
audit log, but raise an explicit exception that lists both of the
affected groups to make it easier to fix manually.