Remove a type: ignore in grouper.ctl
Restructure the conditional to open a file or provide standard input
so that it doesn't require a type: ignore comment.  Move it into the
only file that uses it and make the interface less general, to make
it clearer that nothing else is using this utility function.  (We can
make a proper utility function if there is more than one caller.)