Remove unused cutoff API feature
The API server supported giving an optional cutoff as a parameter
to several routes, which would restrict graph calculations for
group membership and permissions to paths shorter than that cutoff.
This was never supported by Groupy and not tested except as a side
effect of another test, and I see no sign that we ever used it.

Drop this feature to simplify the graph code.

Add type annotations to all changed functions so that mypy has a
better chance of catching errors.  This required cleaning up some
typing in the graph code.