Use the Group entity in the graph
Rather than using a custom GroupTuple type in the graph, reuse the
Group entity from grouper.entities.group.  This required adding the
is_role_user flag to that entity and populating it in the repository,
and removing the uses of some of the other elements of GroupTuple
that weren't strictly necessary.