Switch to PermissionGrant in the graph
Rather than using a custom MappedPermission named tuple to store
permission grants, use the standard PermissionGrant entity.  This
requires splitting permission grants from permission metadata,
which now contains only information about the permission itself,
and reworking how the graph checks whether permissions are audited.

Rename the name field of PermissionGrant to permission, which is
more consistent with the other versions of this entity and makes
for more obvious code in the graph.

Add a test for granting and revoking service account permissions to
ensure that we're exercising the template code that creates links
based on the permission grant ID.