Unify some duplicate graph data structures
Add new private _groups and _permissions dicts that map group and
permission names to Group and Permission objects, and use them to
satisfy requests for the groups and permissions properties on the
fly.  Replace group_tuples, permission_tuples, and permission_metadata
with those new data structures.

Replace the separate users data structure with a property that
generates the list on demand from user_metadata.

Return name and email address metadata about a group as part of
get_group_details, and include email_address in the Group data transfer
object to support this.  This eliminates the separate group_metadata
data structure.

Mark more graph data structures that are not used outside the graph
code as private to make the API cleaner and easier to understand.

These changes eliminate three more database queries from the graph
construction code.