Fix role user group metadata
A previous graph refactoring broke the determination of whether a
group is a role user, and we didn't have a test for that behavior.
Fix the bug and add an explicit test.

Explicitly pass the new user_metadata into get_groups instead of
using self.user_metadata, since the latter is from the previous
version of the graph and may be inconsistent.