Get rid of PermissionGrant
The distinction between PermissionGrant and choosing one of
GroupPermissionGrant and ServiceAccountPermissionGrant was
confusing and error-prone, and we weren't gaining that much from
the two representations.  Use GroupPermissionGrant and
ServiceAccountPermissionGrant everywhere and adjust the code that
creates those objects to gather the additional required information.

Make the mapping_id parameter optional since there are some cases
where we truly don't have a unique ID for the grant, and rename it
to grant_id, which is a more accurate name for what it is.