Use case to retrieve all grants
Provide a use case to retrieve all grants of a permission,
structured as a mapping from permission name to users and service
accounts granted that permission to sets of argument strings from
all of their grants.  This includes modifications to the graph to
cache this data.

This use case is tested but not yet used by a UI.  (It will be used
by the API server in a follow-up diff.)