Add API for retrieving permission grants
/grants will return all permission grants in a JSON dict with the
permissions as keys, followed by users or service_accounts, followed
by the name of the user or service account, followed by a list of
arguments.

/grants/<permission> will return the same for a specific permission,
without the first layer of dict but with a separate top-level key of
permission, indicating the requested permission.

The test uses the low-level API directly since Groupy doesn't have
high-level support for grants.  (It's not clear yet whether it's
worth the trouble of adding that.)