Add usecase to list all enabled users
This use case returns an intermediate level of detail for users,
which matches the detail that we want to provide from the API
`/user-metadata` endpoint: users with their metadata and public
keys.  Don't include passwords, since ideally we wouldn't be
exposing those hashes as readily.  Don't include all group and
permission membership, since that's more expensive to gather and
isn't needed for this use case.

The entities added here won't be right for all future use cases.
A more detailed and possibly a less detailed version will be
added for other use cases in the future, which may require some
renaming.

Duplicate some SQL code in SetupTest for now.  This will go away
once more of that code is pushed into the repository.

This uses direct SQL queries.  The usecase will be implemented on
top of the graph in a later commit before wiring it into the API
server.  There is a forward-looking comment in the repository test
that assumes that has happened and the SQL implementation isn't
directly tested.