Add a user metadata method to the graph
Provide a method on the graph that returns user metadata for all
enabled users using the new transfer objects, and switch the usecase
over to using it by default by creating separate graph and SQL user
repositories and defaulting to the graph repository if the graph is
available.

Have the graph version of the user repository still call through to
the SQL repository for user_is_enabled, since one usecase relies on
disabling a user being immediately reflected by subsequent repository
calls.

Remove the timestamps from user metadata and public keys.  These
were added speculatively, but make the graph implementation very
difficult right now due to different time storage formats.  A later
use case is likely to require these, but we can add them back in when
we cross that bridge; they're not needed for this use case.