Suppress duplicates in permission disable error
When attempting to disable a permission that's still granted to
some groups or service accounts using grouper-ctl, we were showing
the group or service account name for each grant.  This meant that
the same group or service account could be listed multiple times if
it had multiple grants with different arguments.

Since the argument isn't included in the error, this is duplicative
and makes the error harder to read.  De-duplicate the group or
service account names before showing the error.