Separate grants to role users
The ListGrants use case and /grants API endpoint separated users
from service accounts but lumped the deprecated role users in with
users.  This doesn't match our current output needs, so add a
third element for role users and plumb that through to the /grants
route.

This can be deleted later once role users are entirely obsolete.