Handle np-owner in graph cache of permission grants
The new grants cache was not correctly excluding np-owner.  There
were two problems: it was only removing paths with np-owner in the
first edge rather than the last edge, and it wasn't handling the
case where someone is np-owner via a short path but a regular member
via a longer path.

To solve the latter problem, we need to create a new graph that has
only the non-np-owner edges.  It is only used for constructing the
cached grants.