Exclude service accounts from the user part of /grants
We currently have a bug that erroneously allows service accounts to
be added as regular members of groups, causing them to show up in
the user graph.  Work around this by skipping such users based on
user metadata.  This special-case can be removed once we enforce that
the user underlying service accounts cannot be added as a member of
groups.

A better place to put this is to remove service accounts from the
nodes in the graph, but this will break some arguably broken software
that (ab)used the membership of service accounts in groups, so we'll
do that later when we fix that bug.