Fix revocation of service account permissions
When moving the JavaScript for service account permission revocation,
we missed that the URL construction included template variables that
of course are not expanded in the external JavaScript.  Switch to
setting data attributes on the button instead.

This problem was not caught by the test suite despite an explicit
test because the test only checked for a lack of permission rows on
success, and the error page had no permission rows.  Add a positive
test for another page attribute (and add HTML id attributes to more
page elements to make this easier) to confirm that we got back the
service account page rather than some error page.