Set GitHub usernames via oauth to GitHub.
Replace the free-form GitHub username form with a button that runs an OAuth flow with GitHub to determine a users's GitHub username. Both administrators and users can clear a GitHub username from the database with the "Clear" button.

There is now a configuration key for the GitHub APP client id to authenticate to the user. A plugin must be provided to give the GitHub app's secret.

Some deployments may need to talk to the GitHub API through a proxy. Configuration variables for proxy host and port and added to support this. Grouper must also depend on pycurl now, too, because only the curl tornado client supports proxies.