Change the log_exception plugin hook signature
This hook currently isn't used, so before we add an initial use of
it, clean up the types and parameters.  Pass in the full tuple that
Python normally includes with exceptions, in addition to the request
and status if any, and allow the request and status to be omitted
for non-Tornado exceptions.