Convert frontend to new create service account usecase
Move the route definition to stop allowing group IDs instead of
group names, since the usecase world has no good way of mapping a
group ID to a group name.  The support for IDs appears to be
generally unnecessary, so add some comments about moving later
route definitions.

Add some additional information to the error callbacks in the
usecase needed for rendering the page correctly (and fix an
alphabetization error).

Fix the created callback to be outside of the transaction, since
otherwise weird things happen in a Tornado service.