Allow overriding the database source command
We want to stop having every Grouper binary call the same piece of
code to get the database URL so that we can stage rollouts of changes
to the database URL handling as well.  However, given our internal
deployment mechanisms, this means that the path to the command to run
to get the database URL will be different for each component.  To
avoid needing four different config files (eight for stage support),
add a --database-source command-line option to override the
database_source setting in the config and specify a command to run.

grouper-background didn't previously support overriding the database
URL with a command-line flag, so add that.

This doesn't include tests because setting up a proper test of
something that runs an external program is irritatingly difficult,
and I'm bailing on that for now.