Convert service account permission grant to a usecase
Since this now uses the graph properly, it also allows delegating
permissions that are inherited by the group from a parent, rather
than forcing the permission to be granted directly to the group.

I'm doing something a little unusual here in the handler by stashing
some information into the handler object in the post() method that's
then used by the callback methods.  Ideally, everything in the
callback comes from the use case, but in this case the group owner
and the list of permissable permissions to grant are quite strange
to return as part of every error message from the usecase, and (in
the case of the list of grantable permissions) potentially expensive
to recalculate.  This seemed less awkward than forcing the usecase
to calculate and pass them around.