Disable group join button on click
Large groups are having problems with lots of duplicate membership
requests from people clicking on the join submit button multiple
times because they think nothing has happened.  Add JavaScript
support to disable the submit button and switch the cursor to the
progress cursor when the form has been submitted.

It should be possible to reuse the same machinery elsewhere if we
have the same problem.