Don't allow multiple join requests to the same group
If the user already has a pending request to join a group, don't
allow them to request membership again for their user (but still
allow them to request membership for groups they own).  Change the
first prompt, for the member, to be empty if the user themselves
already has a pending request but can still request membership on
behalf of some other group.  Add a warning explaining that this is
the case.