Drop support for Python 2
Remove all uses of six and use the Python 3 form directly.  Remove
remaining documentation and test machinery for Python 2, and document
that Grouper only supports Python 3.

The values() method of an OrderedDict() apparently doesn't return an
iterator, so instead use a temporary list.  (The amount of data should
be small.)

Change the default Travis version of Python to 3.7 to match what we're
using at Dropbox.