Enable chosen search picker for permission names
There are enough permission names that chosen is useful to make
searching easier.  Enable it for permissions, change the field name
to be permission uniformly (not permission_name), and adjust the
Selenium tests as needed.

Change the conventions for the permission request page model in
Selenium to match what we do elsewhere, with separate methods for
every form field.