Reopen the database after frontend tests
In the frontend tests where we verify the correct permission argument
in the database directly, reopen the database first.  Otherwise, the
changes aren't always visible (tends to fail on Travis CI but not in
local testing).

Refactor SetupTest to move the session handling into a new method that
can be called repeatedly.