Don't allow editing the role of a member group
Limit the choices of role for a member that is a group to only
member, and disable the field.  This avoids uncaught exceptions when
someone tries to change the role to something else.

Add mypy typing to the relevant frontend handler and fix a few other
issues.

Use the standard Selenium exception for NoSuchElementException so
there aren't two exceptions with the same name floating around.