Better handle group naming conflicts on rename
There was some code that attempted to handle this by catching
IntegrityError on session commit, but it didn't work since the
IntegrityError was often thrown earlier, and I'm not a fan of
catching generic SQL errors and assuming you know what they mean.
Instead, add an explicit pre-check to see if the new name already
exists, and raise an explicit alert if that's the case.

This adds a race condition if someone creates a group under the
new name in the middle of the submission of the edit form, but we
live with race conditions like that throughout Grouper, and the
worst consequence will be an uncaught exception error page.