Don't allow viewing service accounts under wrong group
Nothing was checking that the group mentioned in the URL when
viewing a service account matched the owner.  Worse, the group from
the URL was used in the page template, so the page would claim the
account was owned by the wrong group.

Verify that the group in the URL matches the owner and return a 404
error if not.