Disallow @ in group names
We cannot remove it from the validation regex because it's used
by role users, and it would make it impossible to view role users.
(We can do this once role users are completely removed.)  Until
then, check for @ specifically in the group creation handler and
reject it.

Add a test for group creation.