Cancel audit warning messages after group disable
If an audited group is disabled, cancel all the audit warning
messages in addition to marking the audit as complete.

When disabling a role user, both mark the audit as complete and
cancel all the audit warning messages.

Add a way to set an environment variable to disable headless mode
when running Selenium tests, since it helps with debugging new
tests.