Allow downgrading your own role
Permit an owner to downgrade their role in a group, but not upgrade
(so np-owner cannot become owner).  Ensure that this doesn't allow
someone to remove the last owner by editing their role.

This redoes the exception handling in Group.add_member() slightly
since otherwise it was spewing errors into the logs when someone
tried to downgrade the role as the last owner.  This isn't an
exceptional case, just a known error, so don't use the noisy
flush_transaction decorator (which is also a little silly here since
AuditLog.log() commits the transaction anyway), and instead manually
wrap the call and roll back if needed.