Use get_path_argument method for all path arguments
Rather than have to remember to unquote the result of kwargs in
every frontend handler, introduce a get_path_argument helper method
and use it uniformly.  This also cleans up some of the typing.

Fix the types and use a more standard style in the GitHub handlers.

Convert PerfProfile to a proper GrouperHandler so that it can use
the same helper.  Excluding it from GrouperHandler seems excessively
nit-picky and it's not being used at the moment anyway.