Test that template dataclasses match template needs
For each template wrapped in a dataclass, test that the dataclass
defines all of the variables that the form needs.

This unfortunately doesn't expand macros, so we have to whitelist
all the macros.  It also doesn't handle inclusion of other files,
so in-line the forms into the templates (which is probably a good
idea anyway).

Rename the permission macro to permission_link, since permission is
used elsewhere as a variable.

Modernize mypy typing of grouper.templating (no modifications were
needed here in the end, but it seemed pointless to throw out the
work).